Fixup usermods_v2_four_line_display_ALT
Get this usermod to build, and restore the interconnect with its
partner usermod_v2_rotary_encoder_ui_ALT.
