+++ usermods/usermod_v2_four_line_display_ALT/library.json
@@ -1,0 +1,7 @@
+{
+  "name:": "four_line_display_ALT",
+  "dependencies": {
+    "U8g2": "~2.34.4",
+    "Wire": ""
+  }  
+}
