--- usermods/usermod_v2_four_line_display_ALT/library.json.disabled
@@ -1,3 +1,0 @@
-{
-  "name:": "usermod_v2_four_line_display_ALT"
-}
