+++ usermods/usermod_v2_four_line_display_ALT/platformio_override.sample.ini
@@ -1,18 +1,11 @@
 [platformio]
-default_envs = esp32dev
+default_envs = esp32dev_fld
 
-[env:esp32dev]
-board = esp32dev
-platform = ${esp32.platform}
-build_unflags = ${common.build_unflags}
+[env:esp32dev_fld]
+extends = env:esp32dev_V4
+custom_usermods = ${env:esp32dev_V4.custom_usermods} four_line_display_ALT
 build_flags =
-    ${common.build_flags_esp32}
-    -D USERMOD_FOUR_LINE_DISPLAY
+    ${env:esp32dev_V4.build_flags}
     -D FLD_TYPE=SH1106
     -D I2CSCLPIN=27
     -D I2CSDAPIN=26
-
-lib_deps =
-    ${esp32.lib_deps}
-    U8g2@~2.34.4
-    Wire
