+++ usermods/usermod_v2_rotary_encoder_ui_ALT/platformio_override.sample.ini
@@ -1,13 +1,11 @@
 [platformio]
-default_envs = esp32dev
+default_envs = esp32dev_re
 
-[env:esp32dev]
-board = esp32dev
-platform = ${esp32.platform}
-build_unflags = ${common.build_unflags}
+[env:esp32dev_re]
+extends = env:esp32dev_V4
+custom_usermods = ${env:esp32dev_V4.custom_usermods} rotary_encoder_ui_ALT
 build_flags =
-    ${common.build_flags_esp32}
-    -D USERMOD_ROTARY_ENCODER_UI
+    ${env:esp32dev_V4.build_flags}
     -D USERMOD_ROTARY_ENCODER_GPIO=INPUT
     -D ENCODER_DT_PIN=21
     -D ENCODER_CLK_PIN=23
